PR feedback: enable go-spew only if Joker built with -tags go_spew
