+++ core/data/core.joke
@@ -4543,7 +4543,11 @@
   (close!__ ch))
 
 (defn- go-spew
-  "Dump ('spew') internal Go structures for object.
+  "Dump ('spew') internal Go structures for object to stderr.
+
+  Returns true if enabled due to building Joker with the 'go_spew' build tag,
+  false otherwise or if some other error occurred (which will be printed to
+  stderr).
 
   Use the optional (map) argument to specify ConfigState
   settings. E.g. {:MaxDepth 10} specifies a maximum depth of 10
@@ -4551,7 +4555,7 @@
 
   For more info, see: https://github.com/jcburley/go-spew/blob/master/spew/config.go"
   {:added "1.0"}
-  (^Nil [o]
+  (^Boolean [o]
    (go-spew__ o))
-  (^Nil [o ^Map cfg]
+  (^Boolean [o ^Map cfg]
    (go-spew__ o cfg)))
